+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Rename BooleanFilter's optionInfo property to just options
 * Drop configured initialMaxRecords for favourited toots to half of the other configuration because it seems to be the bottleneck
 
 ## v0.60.0